requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.8
django-cloudinary-storage==0.3.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
